requirements.txt
openai
langchain
langchain-openai
python-dotenv
streamlit
python-docx
PyPDF2